Removed unneeded cast to uint
The variable is an unsigned int already.

+++ src/GameLogic/AttackableExtensions.cs
@@ -111,7 +111,7 @@ namespace MUnique.OpenMU.GameLogic
             var shieldBypass = Rand.NextRandomBool(attacker.Attributes[Stats.ShieldBypassChance]);
             if (shieldBypass || defender.Attributes[Stats.CurrentShield] < 1)
             {
-                return new HitInfo((uint)damage, 0, 0);
+                return new HitInfo(damage, 0, 0);
             }
 
             var shieldRatio = 0.90;
